requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.2
beanie==2.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.0
h11==0.16.0
idna==3.11
lazy-model==0.3.0
motor==3.7.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.15.3
python-dotenv==1.0.1
requests==2.32.3
selectolax==0.3.17
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0

# Autenticação JWT
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4

# Cache Redis
redis==5.0.1
aioredis==2.0.1

# Rate Limiting
slowapi==0.1.9

# Retry
tenacity==8.2.3

# Logs Estruturados
structlog==23.3.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# IA para Categorização
openai==1.54.0

# Scheduler para tarefas agendadas
apscheduler==3.10.4
